requirements.txt
asgiref==3.6.0+
Django==4.1.5
djangorestframework==3.14.0
python-decouple==3.7
pytz==2022.7
sqlparse==0.4.3
